Update 1
Now takes 10 measurements of pH probe over 1 second, converts those readings to a pH value, throws away the lowest and highest value of the 10 readings, then averages out the remaining 8 values and reports it.

Next is to apply a calibration curve to the value. Also need to eventually look into temperature affecting the pH as well.